Refactor email_logs entity and repository to use Integer ID; update TwilioMailSender to log email sending details

+++ task_management_system/src/main/java/com/example/task_management_system/repositories/EmailLogsRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.example.task_management_system.entities.EmailLogs;
 
 @Repository
-public interface EmailLogsRepository extends JpaRepository<EmailLogs, Long> {
+public interface EmailLogsRepository extends JpaRepository<EmailLogs, Integer> {
 
 }
